Move FindOnShelf types file into find-on-shelf folder & call it "types"
To follow principle of colocation:
(https://kentcdodds.com/blog/colocation).
