+++ src/components/find-on-shelf/types.ts
@@ -1,0 +1,7 @@
+import { HoldingsV3 } from "../../core/fbs/model";
+import { Manifestation } from "../../core/utils/types/entities";
+
+export type ManifestationHoldings = {
+  manifestation: Manifestation;
+  holding: HoldingsV3;
+}[];
