--- src/core/utils/types/find-on-shelf.ts
@@ -1,7 +1,0 @@
-import { HoldingsV3 } from "../../fbs/model";
-import { Manifestation } from "./entities";
-
-export type ManifestationHoldings = {
-  manifestation: Manifestation;
-  holding: HoldingsV3;
-}[];
